Distinguish private attempt oracle failure stages
